# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Project PM API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks/{taskId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add a comment to a task
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /comments/{id}:
    delete:
      operationId: deleteComment
      tags: [comments]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Admin: Users (auth accounts) ─────────────────────────────────────────
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents (all, or filtered by project)
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: global
          in: query
          description: If true, return only global (non-project) documents
          schema:
            type: boolean
        - name: category
          in: query
          schema:
            type: string
        - name: folderId
          in: query
          description: Filter to documents inside this folder (use 0 for root of the given scope)
          schema:
            type: integer
        - name: taskId
          in: query
          description: Filter to documents attached to a specific task
          schema:
            type: integer
        - name: includeTasks
          in: query
          description: When combined with projectId, also include documents attached to any task within that project.
          schema:
            type: boolean
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: createDocument
      tags: [documents]
      summary: Create a document entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    patch:
      operationId: updateDocument
      tags: [documents]
      summary: Update a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Folders ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Comment:
      type: object
      required: [id, taskId, content, authorName, createdAt]
      properties:
        id:         { type: integer }
        taskId:     { type: integer }
        authorId:   { type: ["integer", "null"] }
        authorName: { type: string }
        content:    { type: string }
        createdAt:  { type: string }
    CommentInput:
      type: object
      required: [content, authorName]
      properties:
        authorId:   { type: integer }
        authorName: { type: string, minLength: 1 }
        content:    { type: string, minLength: 1 }

    # ── Admin User (auth account) ───────────────────────────────────────────
...
    Document:
      type: object
      required: [id, title, category, createdAt]
      properties:
        id:          { type: integer }
        projectId:    { type: ["integer", "null"] }
        projectName:  { type: ["string", "null"] }
        taskId:      { type: ["integer", "null"] }
        folderId:    { type: ["integer", "null"] }
        title:       { type: string }
        description: { type: ["string", "null"] }
        url:         { type: ["string", "null"] }
        notes:       { type: ["string", "null"] }
        category:    { type: string, enum: [spec, plan, permit, vendor, as_built, safety, general] }
        uploadedBy:  { type: ["string", "null"] }
        version:     { type: ["string", "null"] }
        createdAt:   { type: string }
        updatedAt:   { type: string }
    DocumentInput:
      type: object
      required: [title, category]
      properties:
        projectId:    { type: integer }
        taskId:      { type: integer }
        folderId:    { type: integer }
        title:       { type: string, minLength: 1 }
        description: { type: string }
        url:         { type: string }
        notes:       { type: string }
        category:    { type: string, enum: [spec, plan, permit, vendor, as_built, safety, general] }
        uploadedBy:  { type: string }
        version:     { type: string }
    DocumentUpdate:
      type: object
      properties:
        projectId:    { type: integer }
        taskId:      { type: ["integer", "null"] }
        folderId:    { type: ["integer", "null"] }
        title:       { type: string }
        description: { type: string }
        url:         { type: string }
        notes:       { type: string }
        category:    { type: string, enum: [spec, plan, permit, vendor, as_built, safety, general] }
        uploadedBy:  { type: string }
        version:     { type: string }
...